// src/app/services/openapi-client/model/categoryDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CategoryDto { 
    id?: number;
    parentId?: number | null;
    name?: string | null;
    slug?: string | null;
    description?: string | null;
    imageUrl?: string | null;
    displayOrder?: number;
    isActive?: boolean;
    productCount?: number;
    subCategories?: Array<CategoryDto> | null;
    createdAt?: string | null;
}