// src/app/services/openapi-client/model/productDetailDto.ts
/**
 * PaintDepotApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProductImageDto } from './productImageDto';
import { ProductVariantDto } from './productVariantDto';


export interface ProductDetailDto { 
    id?: number;
    categoryId?: number | null;
    categoryName?: string | null;
    brandId?: number | null;
    brandName?: string | null;
    sku?: string | null;
    name?: string | null;
    slug?: string | null;
    shortDescription?: string | null;
    fullDescription?: string | null;
    features?: string | null;
    specifications?: string | null;
    usageInstructions?: string | null;
    warnings?: string | null;
    standardCertification?: string | null;
    coverageAreaPerLiter?: number | null;
    basePrice?: number;
    compareAtPrice?: number | null;
    isFeatured?: boolean;
    isBestseller?: boolean;
    isNewArrival?: boolean;
    isActive?: boolean;
    images?: Array<ProductImageDto> | null;
    variants?: Array<ProductVariantDto> | null;
    averageRating?: number | null;
    reviewCount?: number;
    metaTitle?: string | null;
    metaDescription?: string | null;
    createdAt?: string | null;
    updatedAt?: string | null;
}